799 Extracts components out of the summary cards on the start page
